+++ src/searchdhttp.cpp
@@ -25,7 +25,7 @@
 #include "tracer.h"
 #include "searchdbuddy.h"
 
-static bool g_bLogBadHttpReq = val_from_env ( "MANTICORE_LOG_HTTP_BAD_REQ", false ); // log content of bad http reqests, ruled by this env variable
+static bool g_bLogBadHttpReq = val_from_env ( "MANTICORE_LOG_HTTP_BAD_REQ", false ); // log content of bad http requests, ruled by this env variable
 static bool LOG_LEVEL_HTTP = val_from_env ( "MANTICORE_LOG_HTTP", false ); // verbose logging http processing events, ruled by this env variable
 #define LOG_COMPONENT_HTTP ""
 #define HTTPINFO LOGINFO ( HTTP, HTTP )
@@ -169,10 +169,6 @@ public:
 	}
 };
 
-CharStream_c * CreateBlobStream ( const Str_t & sData )
-{
-	return new BlobStream_c ( sData );
-}
 
 /// stream with known content length and no special massage over socket
 class RawSocketStream_c final : public CharStream_c
@@ -243,7 +239,7 @@ public:
 	}
 };
 
-/// stream with known content length and no special massage over socket
+/// chunked stream - i.e. total content length is unknown
 class ChunkedSocketStream_c final: public CharStream_c
 {
 	AsyncNetInputBuffer_c& m_tIn;
@@ -1846,7 +1842,7 @@ protected:
 	}
 };
 
-// stream for ndjsons
+// stream for lines - each 'Read()' returns single line (lines split by \r or \n)
 class NDJsonStream_c
 {
 	CharStream_c& m_tIn;
@@ -1854,7 +1850,7 @@ class NDJsonStream_c
 	Str_t m_sCurChunk { dEmptyStr };
 	bool m_bDone;
 
-	int m_iJsons = 0;
+	int m_iLines = 0;
 	int m_iReads = 0;
 	int m_iTotallyRead = 0; // not used, but provides data during debug
 
@@ -1866,7 +1862,7 @@ public:
 
 	inline bool Eof() const { return m_bDone;}
 
-	Str_t Read()
+	Str_t ReadLine()
 	{
 		assert ( !m_bDone );
 		while (true)
@@ -1880,16 +1876,16 @@ public:
 				++m_iReads;
 			}
 
-			const char* szJson = m_sCurChunk.first;
-			const char* pEnd = szJson + m_sCurChunk.second;
-			const char* p = szJson;
+			const char* szLine = m_sCurChunk.first;
+			const char* pEnd = szLine + m_sCurChunk.second;
+			const char* p = szLine;
 
 			while ( p < pEnd && *p != '\r' && *p != '\n' )
 				++p;
 
 			if ( p==pEnd )
 			{
-				m_dLastChunk.Append ( szJson, p-szJson );
+				m_dLastChunk.Append ( szLine, p-szLine );
 				m_sCurChunk = dEmptyStr;
 				continue;
 			}
@@ -1901,20 +1897,21 @@ public:
 			Str_t sResult;
 			if ( m_dLastChunk.IsEmpty () )
 			{
-				sResult =  { szJson, p - szJson - 1 };
-				if ( IsEmpty ( sResult ) )
-					continue;
-				++m_iJsons;
-				HTTPINFO << "non-last chunk " << m_iJsons << " " << sResult;;
+				sResult =  { szLine, p - szLine - 1 };
+// that is commented out, as we better will deal with empty strings on parser level instead.
+//				if ( IsEmpty ( sResult ) )
+//					continue;
+				++m_iLines;
+				HTTPINFO << "non-last chunk " << m_iLines << " " << sResult;;
 			}
 			else
 			{
-				m_dLastChunk.Append ( szJson, p - szJson );
+				m_dLastChunk.Append ( szLine, p - szLine );
 				sResult = m_dLastChunk;
 				--sResult.second; // exclude terminating \0
 				m_dLastChunk.Resize ( 0 );
-				++m_iJsons;
-				HTTPINFO << "Last chunk " << m_iJsons << " " << sResult;
+				++m_iLines;
+				HTTPINFO << "Last chunk " << m_iLines << " " << sResult;
 			}
 			return sResult;
 		}
@@ -1923,8 +1920,8 @@ public:
 		m_dLastChunk.Add ( '\0' );
 		m_dLastChunk.Resize ( m_dLastChunk.GetLength() - 1 );
 		Str_t sResult = m_dLastChunk;
-		++m_iJsons;
-		HTTPINFO << "Termination chunk " << m_iJsons << " " << sResult;
+		++m_iLines;
+		HTTPINFO << "Termination chunk " << m_iLines << " " << sResult;
 		return sResult;
 	}
 };
@@ -1952,42 +1949,74 @@ public:
 			return false;
 		}
 
+		JsonObj_c tResults ( true );
+		bool bResult = false;
+		int iCurLine = 0;
+		int iLastTxStartLine = 0;
+
+		auto FinishBulk = [&, this] ( ESphHttpStatus eStatus = SPH_HTTP_STATUS_TOTAL ) {
+			JsonObj_c tRoot;
+			tRoot.AddItem ( "items", tResults );
+			tRoot.AddInt ( "current_line", iCurLine );
+			tRoot.AddInt ( "skipped_lines", iCurLine - iLastTxStartLine );
+			tRoot.AddBool ( "errors", !bResult );
+			tRoot.AddStr ( "error", m_sError.IsEmpty() ? "" : m_sError );
+			if ( eStatus == SPH_HTTP_STATUS_TOTAL )
+				eStatus = bResult ? SPH_HTTP_STATUS_200 : SPH_HTTP_STATUS_500;
+			BuildReply ( tRoot.AsString(), eStatus );
+			HTTPINFO << "inserted  " << iCurLine;
+			return !bResult;
+		};
+
+		auto AddResult = [&tResults] ( const char* szStmt, JsonObj_c& tResult ) {
+			JsonObj_c tItem;
+			tItem.AddItem ( szStmt, tResult );
+			tResults.AddItem ( tItem );
+		};
+
 		if ( m_tSource.Eof() )
-		{
-			BuildReply ( R"({"items":[],"errors":false})", SPH_HTTP_STATUS_200 );
-			return true;
-		}
+			return FinishBulk();
 
 		// originally we execute txn for single index
-		// if there is combo, we fall-back to query-by-query commits
-		JsonObj_c tRoot;
-		JsonObj_c tItems ( true );
+		// if there is combo, we fall back to query-by-query commits
+
 		CSphString sTxnIdx;
 		CSphString sStmt;
-		bool bResult = false;
-		int iInserted = 0;
+
 		while ( !m_tSource.Eof() )
 		{
-			auto tQuery = m_tSource.Read();
+			auto tQuery = m_tSource.ReadLine();
+			++iCurLine;
+
+			DocID_t tDocId = 0;
+			JsonObj_c tResult = JsonNull;
+
 			if ( IsEmpty ( tQuery ) )
+			{
+				if ( session::IsInTrans() )
+				{
+					assert ( !sTxnIdx.IsEmpty() );
+					// empty query finishes current txn
+					bResult = ProcessCommitRollback ( FromStr ( sTxnIdx ), tDocId, tResult, m_sError );
+					AddResult ( "bulk", tResult );
+					if ( !bResult )
+						break;
+					sTxnIdx = "";
+					iLastTxStartLine = iCurLine;
+				}
 				continue;
-			++iInserted;
+			}
+
+			bResult = false;
 			auto& tCrashQuery = GlobalCrashQueryGetRef();
 			tCrashQuery.m_dQuery = { (const BYTE*) tQuery.first, tQuery.second };
 			const char* szStmt = tQuery.first;
 			SqlStmt_t tStmt;
 			tStmt.m_bJson = true;
-			DocID_t tDocId = 0;
+
 			CSphString sQuery;
 			if ( !sphParseJsonStatement ( szStmt, tStmt, sStmt, sQuery, tDocId, m_sError ) )
-			{
-				ReportError ( SPH_HTTP_STATUS_400 );
-				HTTPINFO << "inserted  " << iInserted;
-				return false;
-			}
-
-			JsonObj_c tResult = JsonNull;
-			bResult = false;
+				return FinishBulk ( SPH_HTTP_STATUS_400 );
 
 			if ( sTxnIdx.IsEmpty() )
 			{
@@ -1999,12 +2028,12 @@ public:
 				assert ( !sTxnIdx.IsEmpty() );
 				// we should finish current txn, as we got another index
 				bResult = ProcessCommitRollback ( FromStr ( sTxnIdx ), tDocId, tResult, m_sError );
-				sStmt = "bulk";
-				AddResult ( tItems, sStmt, tResult );
+				AddResult ( "bulk", tResult );
 				if ( !bResult )
 					break;
 				sTxnIdx = tStmt.m_sIndex;
 				ProcessBegin ( sTxnIdx );
+				iLastTxStartLine = iCurLine;
 			}
 
 			switch ( tStmt.m_eStmt )
@@ -2029,17 +2058,19 @@ public:
 				break;
 
 			default:
-				ReportError ( "Unknown statement", SPH_HTTP_STATUS_400 );
-				HTTPINFO << "inserted  " << iInserted;
-				return false;
+				m_sError = "Unknown statement";
+				return FinishBulk ( SPH_HTTP_STATUS_400 );
 			}
 
 			if ( !bResult || !session::IsInTrans() )
-				AddResult ( tItems, sStmt, tResult );
+				AddResult ( sStmt.cstr(), tResult );
 
 			// no further than the first error
 			if ( !bResult )
 				break;
+
+			if ( !session::IsInTrans() )
+				iLastTxStartLine = iCurLine;
 		}
 
 		if ( bResult && session::IsInTrans() )
@@ -2048,27 +2079,16 @@ public:
 			// We're in txn - that is, nothing committed, and we should do it right now
 			JsonObj_c tResult;
 			bResult = ProcessCommitRollback ( FromStr ( sTxnIdx ), 0, tResult, m_sError );
-			sStmt = "bulk";
-			AddResult ( tItems, sStmt, tResult );
+			AddResult ( "bulk", tResult );
+			if ( bResult )
+				iLastTxStartLine = iCurLine;
 		}
 
 		session::SetInTrans ( false );
-
-		tRoot.AddItem ( "items", tItems );
-		tRoot.AddBool ( "errors", !bResult );
-		BuildReply ( tRoot.AsString(), bResult ? SPH_HTTP_STATUS_200 : SPH_HTTP_STATUS_500 );
-		HTTPINFO << "inserted  " << iInserted;
-		return true;
+		return FinishBulk();
 	}
 
-protected:
-	static void AddResult ( JsonObj_c & tRoot, CSphString & sStmt, JsonObj_c & tResult )
-	{
-		JsonObj_c tItem;
-		tItem.AddItem ( sStmt.cstr(), tResult );
-		tRoot.AddItem ( tItem );
-	}
-
+private:
 	const char* IsNDJson() const
 	{
 		if ( !m_tOptions.Exists ( "content-type" ) )